requirements.txt
python-telegram-bot==20.3
transformers==4.40.0
scikit-learn
joblib
